tweak bbdb-site.el.in to bbdb-site.el generation process
Use a more symmetric and robust token for where the real tex directory
belongs in lisp/bbdb-site.el.in

Use "sed -e", which is POSIX, instead of "sed --expression", which ain't.

Add appropriate automake magic to cause bbdb-site.el to be generated
following reconfiguration, to be generated early as it might be
"required" by other lisp source files, and to be included in the
distribution tarball.